refactor: extract ClassicBrowserWOMService from ClassicBrowserService
Extract WOM (Working-Object Memory) logic into dedicated service:
- Move tab-to-object mapping management to WOMService
- Move tab group creation and update logic to WOMService
- Maintain event-driven architecture for loose coupling
- Keep tabGroupId in StateService as single source of truth
- Extract URL helper functions to separate module

This reduces ClassicBrowserService complexity and improves separation of concerns
between view management and data layer interactions.
